======= Squeaky Window (7 kyu) ======= Your function takes in parameter an array of integers nums and a strictly-positive integer k.
Imagine a sliding window of size k which is moving from left to right
in the array.

You can only see the k elements that are in the window.

The sliding window moves right by one position until it is no longer possible
to have k elements in the window.

You have to return an array of the maximums for each successive window,
from left to right.