====== Missing Angle (6 kyu) ====== Below is a right-angled triangle:
  |  |   |    |   \
o |    \ h
  |       |    θ   |_______\
     a

Your challange is to write a function that calculates the angle θ in degrees to
the nearest integer.

You will be given three arguments representing each side: o, h and a.

One of the arguments equals zero.

Use the length of the two other sides to calculate θ.

You will not be expected to handle any erronous data in your solution.