======== Contamination #1-String- (8 kyu) ======== An AI has infected a text with a charactergit add contamination.js
This text is now fully mutated to this character.

If the text or the character are empty, return an empty string.
There will never be a case when both are empty as nothing is going ongit add contamination.js

Note: The character is a string of length 1 or an empty string.
Example

text before = 'abc'
character   = 'z'
text after  = 'zzz'